+++ lib/components/fo_dropdown_list/fo_dropdown_list_component.dart
@@ -41,6 +41,9 @@ class FoDropdownListComponent<T> implements AfterChanges, OnDestroy {
   bool constrainToViewPort = true;
 
   @Input()
+  bool showCategoryLabels = true;
+
+  @Input()
   bool materialIcons = true;
 
   @Input()
@@ -104,8 +107,7 @@ class FoDropdownListComponent<T> implements AfterChanges, OnDestroy {
     _selectController.close();
   }
 
-  void onSelect(html.Event e, FoDropdownOptionRenderable option) {
-    e.stopPropagation();
+  void onSelect(html.Event event, FoDropdownOptionRenderable option) {
     _selectController.add(option);
   }