+++ lib/components/fo_text_input/fo_text_input_component.dart
@@ -59,6 +59,12 @@ class FoTextInputComponent
   Map<String, List<FoDropdownOptionRenderable>> options;
 
   @Input()
+  bool showDropdownCategoryLabels = true;
+
+  @Input()
+  bool showDropdownOnFocus = false;
+
+  @Input()
   bool disabled = false;
 
   /// Make sure options dropdown doesn't extend beyond the viewport
@@ -88,13 +94,22 @@ class FoTextInputComponent
   final StreamController<html.FocusEvent> _focusController =
       StreamController<html.FocusEvent>();
   final StreamController<String> _blurController = StreamController<String>();
-  bool dropdownVisible = false;
+  bool _dropdownVisible = false;
   html.Element host;
   html.InputElement inputElement;
   bool hasFocus = false;
   FoTextInputComponent(
       @Self() @Optional() this.control, this.host, this._changeDetectorRef) {
     if (control != null) control.valueAccessor = this;
+  }
+
+  bool get dropdownVisible =>
+      options != null &&
+      (_dropdownVisible && value?.isEmpty == false ||
+          (hasFocus == true && showDropdownOnFocus == true));
+
+  set dropdownVisible(bool value) {
+    _dropdownVisible = value;
   }
 
   /// Action button triggered
@@ -177,7 +192,7 @@ class FoTextInputComponent
     // Prevent the input from gaining focus
     event.preventDefault();
     value = '';
-    dropdownVisible = false;
+    _dropdownVisible = false;
     if (_onChange != null) {
       _onChange(value);
     }
@@ -191,7 +206,8 @@ class FoTextInputComponent
 
   void onFilterSelect(FoDropdownOptionRenderable event) {
     value = event.renderLabel;
-    dropdownVisible = false;
+    _dropdownVisible = false;
+
     if (_onChange != null) {
       _onChange(value);
     }
@@ -199,6 +215,8 @@ class FoTextInputComponent
   }
 
   void onFocus(html.FocusEvent event) {
+    event..preventDefault();
+
     hasFocus = true;
     _focusController.add(event);
   }
@@ -218,7 +236,7 @@ class FoTextInputComponent
       _onChange(value);
     }
 
-    dropdownVisible = options != null && value?.isEmpty == false;
+    _dropdownVisible = options != null && value?.isEmpty == false;
   }
 
   @override